% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_raster.r
\name{dist_raster}
\alias{dist_raster}
\title{Distance raster function}
\usage{
dist_raster(poly, r, field = NULL)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
