% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_raster.r
\name{dummy_raster}
\alias{dummy_raster}
\title{Dummy raster function}
\usage{
dummy_raster(poly, r, field)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
