% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pta_to_raster.r
\name{pta_to_raster}
\alias{pta_to_raster}
\title{PTA to raster function}
\usage{
pta_to_raster(pta, crs = NULL, res = c(1000, 1000))
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
