% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_raster.r
\name{quantile_raster}
\alias{quantile_raster}
\title{Quantile raster function}
\usage{
quantile_raster(rst, sep.cdf = FALSE)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
