% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_cross_cor.r
\name{raster_cross_cor}
\alias{raster_cross_cor}
\title{Raster cross-correlation function}
\usage{
raster_cross_cor(x, y, lag = 0, invar = NA)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
