% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_line_coverage.r
\name{raster_line_coverage}
\alias{raster_line_coverage}
\title{Line coverage on raster function}
\usage{
raster_line_coverage(
  line,
  rst,
  res.fact = 5,
  mask = T,
  prec = 1e+05,
  binary = TRUE,
  is_line = TRUE
)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
