% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_line_length.r
\name{raster_line_length}
\alias{raster_line_length}
\title{Line length on raster function}
\usage{
raster_line_length(line, rst, mask = T, prec = 1e+05)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
