% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_point_count.r
\name{raster_point_count}
\alias{raster_point_count}
\title{Raster point count function}
\usage{
raster_point_count(point, rst, year.col = NULL, prefix = NULL)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
