% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_point_id.r
\name{raster_point_id}
\alias{raster_point_id}
\title{Number of point within raster function}
\usage{
raster_point_id(point, rst, id.col = NULL, year.col = NULL, prefix = NULL)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
