% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_time_coverage.r
\name{raster_time_coverage}
\alias{raster_time_coverage}
\title{Time raster function}
\usage{
raster_time_coverage(
  line,
  type = c("month", "week", "yday"),
  date.col,
  rst,
  mask = T,
  prec = 1e+05
)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
