% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_stretch.r
\name{rng_stretch}
\alias{rng_stretch}
\title{Linearly stretch function}
\usage{
rng_stretch(x, lb = 0, ub = 1)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
