% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpool_scale.r
\name{rpool_scale}
\alias{rpool_scale}
\title{Recale raster by pooling the stack function}
\usage{
rpool_scale(rst)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
dist_raster(poly, r, field = NULL)
}
\keyword{last}
