% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_rst.r
\name{stack_rst}
\alias{stack_rst}
\title{stack raster function}
\usage{
stack_rst(r.list, rst)
}
\arguments{
\item{v}{A vector.}
}
\description{
This function allows you to extract the las element of a vector.
}
\examples{
y<-rnorm(10,0,1)
stack(r.list, rst)
}
\keyword{last}
